Refactor scripts and consolidate parser
