+++ index.html
@@ -3,6 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <title>Timesheet Parser MVP</title>
+  <link rel="stylesheet" href="./style.css">
 </head>
 <body>
 
@@ -43,7 +44,7 @@
     <button onclick="addRule()">Add New Rule</button>
   </section>
 
-  <script type="module" src="/app.js"></script>
+  <script type="module" src="./src/main.js"></script>
 </body>
 
 </html>
